hide import feature behind feature gate
since this feature isn’t ready to release yet let’s put it behind a
feature gate so we can control who sees it.

notice we only put the ui entry point behind the feature gate. we don’t
want to add too much complexity to protect something from users that
they’ll soon be getting anyway. the goal is to do whatever is necessary
to hide the entry points but keep the feature gate as simple to remove
as possible to avoid complications when it’s ready to be released to
the world.