process the import file
we can continue to make commits to the import feature while it is
behind the feature gate.

notice we’re not using a feature gate in the job. since the ui is
hidden we don’t worry about this running unintentionally. of course
there could sometimes be good reasons to check the feature gate in
multiple places, ex: you have differing behavior in more than one place
rather than a complete new code branch.